Adjustment so loadFile can load saved file
On Android, to load a saved file, it must start with "file://".  In previous versions of Cocos fullPathForFilename() returned the entered filename on failure, so before I could pass in a file path that already had "file://" on the front.  Now fullPathForFilename() returns "" when it fails to find a file, so I must pass in the non URL version path that starts with "/data/data" to get it to pass through fullPathForFilename(), but then before it can loadFile it must add "file://" to the start of the path for me.